ingestor/requirements.txt
mysql-connector-python
boto3
chalice
